% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pascal_matrix.R
\name{generate_symm_pascal_matrix}
\alias{generate_symm_pascal_matrix}
\title{Generate a symmetric pascal matrix}
\usage{
generate_symm_pascal_matrix(n)
}
\arguments{
\item{n}{a positive integer, number of rows and columns}
}
\value{
a nxn symmetric pascal matrix
}
\description{
Generate a symmetric pascal matrix
}
