% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pascal_matrix.R
\name{generate_tri_pascal_matrix}
\alias{generate_tri_pascal_matrix}
\title{Generate Pascal's triangle in matrix form}
\usage{
generate_tri_pascal_matrix(n, position = "lower")
}
\arguments{
\item{n}{a positive integer, number of rows and columns}

\item{position}{either "lower" for an lower-triangular matrix, the default, or "upper" for an upper-triangular matrix}
}
\value{
a nxn Pascal matrix
}
\description{
Generate Pascal's triangle in matrix form
}
