% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pascal_matrix.R
\name{pascal_matrix_plot}
\alias{pascal_matrix_plot}
\title{Plot Sierpiński triangle based on pascal matrix}
\usage{
pascal_matrix_plot(m)
}
\arguments{
\item{m}{a pascal matrix}
}
\value{
a plot with Sierpiński triangle
}
\description{
Plot Sierpiński triangle based on pascal matrix
}
\examples{
m <- generate_tri_pascal_matrix(10)
plot <- pascal_matrix_plot(m)
plot
}
